requirements.txt
fastmcp==2.10.5
mcp==1.11.0
openai==1.95.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
